Refactored Log4j2ServiceNameInstrumentation and Log4j2ServiceVersionInstrumentation
